requirements.txt
pytorch
torchvision
tqdm
torchsummary